Rename app identifiers to com.keplr.vizor
