+++ android/app/src/main/kotlin/com/keplr/vizor/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.zcash.zcash_wallet
+package com.keplr.vizor
 
 import io.flutter.embedding.android.FlutterActivity
