+++ ios/Runner/Info.plist
@@ -9,8 +9,8 @@
 	</array>
 	<key>BGTaskSchedulerPermittedIdentifiers</key>
 	<array>
-		<string>com.zcash.zcashWallet.sync</string>
-		<string>com.zcash.zcashWallet.txtrack</string>
+			<string>com.keplr.vizor.sync</string>
+			<string>com.keplr.vizor.txtrack</string>
 		<string>com.pravera.flutter_foreground_task.refresh</string>
 	</array>
 	<key>NSSupportsLiveActivities</key>
