+++ macos/Runner/Configs/AppInfo.xcconfig
@@ -8,7 +8,7 @@
 PRODUCT_NAME = zcash_wallet
 
 // The application's bundle identifier
-PRODUCT_BUNDLE_IDENTIFIER = com.zcash.zcashWallet
+PRODUCT_BUNDLE_IDENTIFIER = com.keplr.vizor
 
 // The copyright displayed in application information
 PRODUCT_COPYRIGHT = Copyright © 2026 com.zcash. All rights reserved.
